% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterize.R
\name{rasterize}
\alias{rasterize}
\title{Bin environmental data}
\usage{
rasterize(X, n = 10, precisions = NULL)
}
\arguments{
\item{X}{matrix (or data.frame) of environmental data.}

\item{n}{number of bins for each variable in the data (just one number).}

\item{precisions}{the precision used to cut each column of the data into bins (a vector, as long as the number of columns of X); overrides \code{n}.}
}
\value{
A data.frame with all observed combinations of binned variables and a column \code{n} containing the number of observations of this combination. Each bin is identified by the average value of each variable in the bin. Therefore, it changes from bin to bin, i.e. even with a precision of 0.1, binned values are not all 0.05, 0.15, etc.
}
\description{
Bin each environment variable and count the number of observation of all combinations of bins. Is then used to filter out rare combinations of variables.
}
\examples{
X <- data.frame(a=runif(100), b=runif(100)*10)
rasterize(X, n=2)
rasterize(X, precisions=c(0.5, 5))
}

